Add 8b8tCore Plugin and 8b8t server to readme
Hello, 
We would like 8b8tCore to be added to the README file as a list of supported Folia plugins. Even though we may have our issues and differences, this does not mean we should be discriminated against and we should have this useful utility plugin that many developers use on their server that fits their needs.